Splits tests into fast unit tests and slow functional tests
